fix: widen Note::time() return type from i32 to i64 to prevent overflow at ~35.8 minutes
Note::time() cast i64 microseconds to i32 milliseconds, causing overflow
for songs longer than ~35.8 minutes (i32::MAX ms). This matches the fix
already applied to TimeLine::time() which returns i64.

- Change Note::time() to return i64, removing the narrowing `as i32` cast
- Change Lane::mark() parameter from i32 to i64 to match
- Remove now-unnecessary `as i64` casts on note.time() call sites

+++ crates/bms-model/src/lane.rs
@@ -67,7 +67,7 @@ impl Lane {
         self.hiddenseekpos = self.hiddenbasepos;
     }
 
-    pub fn mark(&mut self, time: i32) {
+    pub fn mark(&mut self, time: i64) {
         if !self.notes.is_empty() {
             while self.notebasepos < self.notes.len() - 1
                 && self.notes[self.notebasepos + 1].time() < time
